app/requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.1
psycopg2-binary==2.9.11
redis==7.4.0
sqlalchemy==2.0.48
celery==5.6.3
flask-migrate==4.1.0
flask-celery==2.4.3
gunicorn==25.3.0